Hoisting test out of server
