+++ src/test/cli/metrics/issue-count-test.ts
