+++ src/test/cli/metrics/stars-test.ts
